Move `addDestroyEffects` and `withCorrectTint` to dynamic client hooks
